chore: rename icons resolver module
